requirements.txt
Flask==2.3.0
requests==2.28.2
yfinance==0.2.25
python-dotenv==1.0.0
